requirements.txt
requests
Pillow
RPi.GPIO
spidev
pytest